MacGui; fixed warnings related to 64bit cocoa.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@2183 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/ChapterTitles.h
@@ -19,10 +19,10 @@
 
 - (id)tableView:(NSTableView *)aTableView
       objectValueForTableColumn:(NSTableColumn *)aTableColumn
-      row:(int)rowIndex;
+      row:(NSInteger)rowIndex;
       
 - (void)tableView:(NSTableView *)aTableView
         setObjectValue:(id)anObject
         forTableColumn:(NSTableColumn *)aTableColumn
-        row:(int)rowIndex;
+        row:(NSInteger)rowIndex;
 @end
